requirements.txt
# Core
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic-settings==2.2.0

# AI & chat
google-generativeai==0.6.0
openai==1.7.0

# Database / cache
redis==5.3.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.0
httpx==1.3.0
